move CHC results to Library-CHC
