+++ trunk/source/Library-CHC/META-INF/MANIFEST.MF
@@ -14,5 +14,6 @@ Require-Bundle: de.uni_freiburg.informatik.ultimate.lib.smtlibutils,
  de.uni_freiburg.informatik.ultimate.lib.automata,
  de.uni_freiburg.informatik.ultimate.lib.boogieast
 Export-Package: de.uni_freiburg.informatik.ultimate.lib.chc,
- de.uni_freiburg.informatik.ultimate.lib.chc.eldarica
+ de.uni_freiburg.informatik.ultimate.lib.chc.eldarica,
+ de.uni_freiburg.informatik.ultimate.lib.chc.results
 Automatic-Module-Name: de.uni.freiburg.informatik.ultimate.lib.chc
