+++ trunk/source/Library-CHC/src/de/uni_freiburg/informatik/ultimate/lib/chc/results/ChcSatResult.java
@@ -4,46 +4,45 @@
  * Copyright (C) 2014-2015 Matthias Heizmann (heizmann@informatik.uni-freiburg.de)
  * Copyright (C) 2015 University of Freiburg
  *
- * This file is part of the ULTIMATE Core.
+ * This file is part of the ULTIMATE CHC Library.
  *
- * The ULTIMATE Core is free software: you can redistribute it and/or modify
+ * The ULTIMATE CHC Library is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published
  * by the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * The ULTIMATE Core is distributed in the hope that it will be useful,
+ * The ULTIMATE CHC Library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Lesser General Public License for more details.
  *
  * You should have received a copy of the GNU Lesser General Public License
- * along with the ULTIMATE Core. If not, see <http://www.gnu.org/licenses/>.
+ * along with the ULTIMATE CHC Library. If not, see <http://www.gnu.org/licenses/>.
  *
  * Additional permission under GNU GPL version 3 section 7:
- * If you modify the ULTIMATE Core, or any covered work, by linking
+ * If you modify the ULTIMATE CHC Library, or any covered work, by linking
  * or combining it with Eclipse RCP (or a modified version of Eclipse RCP),
  * containing parts covered by the terms of the Eclipse Public License, the
- * licensors of the ULTIMATE Core grant you additional permission
+ * licensors of the ULTIMATE CHC Library grant you additional permission
  * to convey the resulting work.
  */
-package de.uni_freiburg.informatik.ultimate.core.lib.results;
+package de.uni_freiburg.informatik.ultimate.lib.chc.results;
 
 import de.uni_freiburg.informatik.ultimate.core.model.results.IResult;
 
 /**
- * If TreeAutomizer finds out "SAT", it will report this result.
- * In the future this may contain a model.
- * 
+ * If TreeAutomizer finds out "SAT", it will report this result. In the future this may contain a model.
+ *
  * @author Alexander Nutz (nutz@informatik.uni-freiburg.de)
  *
  */
 public class ChcSatResult implements IResult {
 
-	private String mLongDescription;
-	private String mShortDescription;
-	private String mPlugin;
+	private final String mLongDescription;
+	private final String mShortDescription;
+	private final String mPlugin;
 
-	public ChcSatResult(String plugin, String shortDescription, String longDescription) {
+	public ChcSatResult(final String plugin, final String shortDescription, final String longDescription) {
 		mPlugin = plugin;
 		mShortDescription = shortDescription;
 		mLongDescription = longDescription;
