+++ trunk/source/Library-UltimateTest/META-INF/MANIFEST.MF
@@ -21,6 +21,7 @@ Require-Bundle: de.uni_freiburg.informatik.ultimate.lib.core,
  de.uni_freiburg.informatik.ultimate.core,
  de.uni_freiburg.informatik.ultimate.lib.util,
  org.junit,
+ de.uni_freiburg.informatik.ultimate.lib.chc,
  de.uni_freiburg.informatik.ultimate.lib.smtlib,
  de.uni_freiburg.informatik.ultimate.smtsolver.external,
  de.uni_freiburg.informatik.ultimate.smtinterpol,
